refactor(openomni): extract shared message helpers from subagent runtime
